Use AT's real-time board data
The Maxx "API" was pretty flaky. The AT "API" seems much more stable.

Sadly, while AT has a public API, it's completely useless and
only has fake data for realtime boards. Their
"public-restricted" API has the data, but is not documented and
does not seem to accept API keys from their public site.

This makes me sad, but the app is much more responsive and
reliable with this API underneath it.